Route constants moved from route folder to constants folder
