+++ src/component/Menu.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import { Link } from 'react-router-dom';
 import '../App.css';
-import ROUTES from '../routes/routes';
+import ROUTES from '../constants/routes';
 
 /**
  * This class contains Menu Elements.
